Refactor: 경로 수정 및 Stories 폴더 생성
